+++ docs/conf.py
@@ -349,9 +349,11 @@ nitpick_ignore = [
     ("py:class", "CommandResult"),
     ("py:class", "CommunicationCallback"),
     ("py:class", "DynamicProxy"),
+    ("py:class", "IActivityGateway"),
     ("py:class", "IClientDispatcher"),
     ("py:class", "IDatabaseDispatcher"),
     ("py:class", "IDatasetGateway"),
+    ("py:class", "IPlanGateway"),
     ("py:class", "LocalClient"),
     ("py:class", "OID_TYPE"),
     ("py:class", "Path"),
